+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # bract.ring Change Log
 
+## TODO
+
+None
+
+
+## [WIP] 0.4.0 / 2017-August-??
+
+- Use the GA version of bract.core 0.4.0
+
+
+## 0.4.0-alpha2 / 2017-August-01
+
+- Use bract.core 0.4.0-alpha2
+
+
+## 0.4.0-alpha1 / 2017-July-31
+
+- Use bract.core 0.4.0-alpha1
+- [BREAKING CHANGE] Rename `bract.ring.config` namespace to `bract.ring.keydef`
+
+
 ## 0.3.1 / 2017-June-30
 - Use bract.core 0.3.1
 - Update inducer docstring to reflect the context keys they use